Expand the reindexing capabilities.
We can now move to a multi step reindex. This allows us to update the
index alias at a later stage if we choose.
It also allows for updating an index from another index without creating
it. This allows partial updates using version_type=external.
Also set the number of replicas to 0 when bulk-indexing is set.